# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Goster IoT 管理 API
  version: 0.4.0
  summary: 面向当前已实现管理能力的前后端契约。
  description: |
    本契约基于当前系统行为与数据模型整理。
    它定义了稳定的 `/api/v1` 命名空间用于前后端解耦，
    同时保留现有管理逻辑的兼容性预期。

    设计原则：
    - 契约优先
    - 向后安全演进
    - 通过 `extensions` 与 `meta` 扩展负载
  x-current-logic:
    metrics_default_range: 1h
    metrics_min_valid_timestamp_ms: 1672531200000
    auth_primary_mode: cookie_session

servers:
  - url: http://localhost:8080
    description: 本地开发环境

tags:
  - name: Auth
  - name: Device
  - name: Metrics
  - name: User

security:
  - CookieSession: []
  - BearerAuth: []

paths:
  /api/v1/auth/captcha/config:
    get:
      tags: [Auth]
      operationId: getCaptchaConfig
      summary: 获取前端验证码服务配置。
      security: []
      responses:
        '200':
          description: 验证码配置。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptchaConfigResponse'

  /api/v1/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: 注册用户账号。
      description: |
        与现有注册行为保持一致：
        - 首个注册用户获得 `Admin` 权限。
        - 后续用户默认获得 `None` 权限。
      x-legacy-endpoint: /auth/register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 登录并创建认证会话。
      x-legacy-endpoint: /auth/login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功。
          headers:
            Set-Cookie:
              schema:
                type: string
              description: 由服务端设置的会话 Cookie。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 登出当前会话。
      x-legacy-endpoint: /auth/logout
      responses:
        '204':
          description: 已登出。
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: 获取当前已认证用户信息。
      responses:
        '200':
          description: 当前用户信息。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/devices:
    get:
      tags: [Device]
      operationId: listDevices
      summary: 按认证状态列出设备。
      parameters:
        - $ref: '#/components/parameters/DeviceStatusFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 设备列表。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/devices/{uuid}:
    get:
      tags: [Device]
      operationId: getDevice
      summary: 获取设备详情。
      parameters:
        - $ref: '#/components/parameters/DeviceUUID'
      responses:
        '200':
          description: 设备详情。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Device]
      operationId: deleteDevice
      summary: 永久删除设备及其关联数据。
      parameters:
        - $ref: '#/components/parameters/DeviceUUID'
      responses:
        '204':
          description: 已删除。
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/devices/{uuid}/approve:
    post:
      tags: [Device]
      operationId: approveDevice
      summary: 审批待认证设备。
      x-legacy-endpoint: /device/approve?uuid={uuid}
      parameters:
        - $ref: '#/components/parameters/DeviceUUID'
      responses:
        '200':
          description: 已通过审批。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/devices/{uuid}/revoke:
    post:
      tags: [Device]
      operationId: revokeDevice
      summary: 拒绝待认证设备或撤销已认证设备。
      x-legacy-endpoint: /device/revoke?uuid={uuid}
      parameters:
        - $ref: '#/components/parameters/DeviceUUID'
      responses:
        '200':
          description: 已撤销/已拒绝。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/devices/{uuid}/unblock:
    post:
      tags: [Device]
      operationId: unblockDevice
      summary: 将设备从已拒绝/已撤销状态移回待认证状态。
      x-legacy-endpoint: /device/unblock?uuid={uuid}
      parameters:
        - $ref: '#/components/parameters/DeviceUUID'
      responses:
        '200':
          description: 已解除阻断。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/devices/{uuid}/token/refresh:
    post:
      tags: [Device]
      operationId: refreshDeviceToken
      summary: 刷新设备令牌。
      x-legacy-endpoint: /device/token/refresh?uuid={uuid}
      parameters:
        - $ref: '#/components/parameters/DeviceUUID'
      responses:
        '200':
          description: 令牌已刷新。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/metrics/{uuid}:
    get:
      tags: [Metrics]
      operationId: getMetrics
      summary: 查询设备指标点数据。
      x-legacy-endpoint: /api/metrics/{uuid}
      parameters:
        - $ref: '#/components/parameters/DeviceUUID'
        - $ref: '#/components/parameters/MetricsRange'
        - $ref: '#/components/parameters/StartMs'
        - $ref: '#/components/parameters/EndMs'
      responses:
        '200':
          description: 指标点数据。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/users:
    get:
      tags: [User]
      operationId: listUsers
      summary: 列出系统用户。
      responses:
        '200':
          description: 用户列表。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/users/{username}/permission:
    post:
      tags: [User]
      operationId: updateUserPermission
      summary: 更新用户权限等级。
      x-legacy-endpoint: /user/permission
      parameters:
        - $ref: '#/components/parameters/Username'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePermissionRequest'
      responses:
        '200':
          description: 权限已更新。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    CookieSession:
      type: apiKey
      in: cookie
      name: goster_session
      description: 当前主要认证方式。
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 为未来基于令牌的认证预留。

  parameters:
    DeviceUUID:
      name: uuid
      in: path
      required: true
      schema:
        type: string
        minLength: 1

    Username:
      name: username
      in: path
      required: true
      schema:
        type: string
        minLength: 1

    DeviceStatusFilter:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [all, authenticated, pending, refused, revoked]
        default: authenticated
      description: |
        认证状态筛选：
        - authenticated -> `AuthenticateStatus=0`
        - refused -> `AuthenticateStatus=1`
        - pending -> `AuthenticateStatus=2`
        - revoked -> `AuthenticateStatus=4`
        - all -> 不筛选

    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1

    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100

    MetricsRange:
      name: range
      in: query
      required: false
      schema:
        type: string
        enum: [1h, 6h, 24h, 7d, all]
        default: 1h

    StartMs:
      name: start_ms
      in: query
      required: false
      schema:
        type: integer
        format: int64
      description: 显式指定开始时间戳（毫秒）。与 `end_ms` 同时提供时会覆盖 `range`。

    EndMs:
      name: end_ms
      in: query
      required: false
      schema:
        type: integer
        format: int64
      description: 显式指定结束时间戳（毫秒）。与 `start_ms` 同时提供时会覆盖 `range`。

  responses:
    BadRequest:
      description: 无效请求。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: 需要认证或会话已过期。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: 权限不足。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: 资源不存在。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Conflict:
      description: 资源已存在或状态冲突。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ApiResponseBase:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: integer
          description: 0 表示成功，非 0 表示错误。
          example: 0
        message:
          type: string
          example: ok
        request_id:
          type: string
          description: 用于链路追踪的关联 ID。
          example: req_20260312_abcdef
        meta:
          type: object
          additionalProperties: true
          description: 可扩展元数据。

    ErrorDetail:
      type: object
      required: [type]
      properties:
        type:
          type: string
          example: validation_error
        field:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true

    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [error]
          properties:
            data:
              type: 'null'
            error:
              $ref: '#/components/schemas/ErrorDetail'

    PermissionType:
      type: integer
      enum: [0, 1, 2, 3]
      description: |
        0=无权限, 1=只读, 2=读写, 3=管理员
      x-enumNames: [None, ReadOnly, ReadWrite, Admin]

    AuthStatusType:
      type: integer
      enum: [0, 1, 2, 3, 4]
      description: |
        0=已认证, 1=已拒绝, 2=待认证, 3=未知, 4=已撤销
      x-enumNames: [Authenticated, Refused, Pending, Unknown, Revoked]

    DeviceRuntimeStatus:
      type: integer
      enum: [0, 1, 2]
      description: 0=离线, 1=在线, 2=延迟
      x-enumNames: [Offline, Online, Delayed]

    MetricType:
      type: integer
      enum: [1, 2, 4]
      description: 1=温度, 2=湿度, 4=光照
      x-enumNames: [Temperature, Humidity, Lux]

    CaptchaConfig:
      type: object
      required: [provider, enabled]
      properties:
        provider:
          type: string
          enum: [none, turnstile]
        enabled:
          type: boolean
        site_key:
          type: string
          nullable: true
        extensions:
          type: object
          additionalProperties: true

    CaptchaConfigResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/CaptchaConfig'

    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 8
          maxLength: 128
        email:
          type: string
          format: email
          nullable: true
        captcha_token:
          type: string
          nullable: true
        extensions:
          type: object
          additionalProperties: true

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        remember_me:
          type: boolean
          default: false
        captcha_token:
          type: string
          nullable: true
        extensions:
          type: object
          additionalProperties: true

    AuthSession:
      type: object
      required: [username, permission]
      properties:
        username:
          type: string
        email:
          type: string
          nullable: true
        permission:
          $ref: '#/components/schemas/PermissionType'
        authenticated:
          type: boolean
          default: true
        extensions:
          type: object
          additionalProperties: true

    AuthSessionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/AuthSession'

    DeviceMetadata:
      type: object
      required: [name, hw_version, sw_version, config_version, sn, mac, created_at, authenticate_status]
      properties:
        name:
          type: string
        hw_version:
          type: string
        sw_version:
          type: string
        config_version:
          type: string
        sn:
          type: string
        mac:
          type: string
        created_at:
          type: string
          format: date-time
        token:
          type: string
          nullable: true
        authenticate_status:
          $ref: '#/components/schemas/AuthStatusType'
        extensions:
          type: object
          additionalProperties: true

    DeviceRuntime:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/DeviceRuntimeStatus'
        status_text:
          type: string
          enum: [offline, online, delayed]
        last_seen_ms:
          type: integer
          format: int64
          nullable: true
        extensions:
          type: object
          additionalProperties: true

    DeviceRecord:
      type: object
      required: [uuid, meta]
      properties:
        uuid:
          type: string
        meta:
          $ref: '#/components/schemas/DeviceMetadata'
        runtime:
          $ref: '#/components/schemas/DeviceRuntime'
        extensions:
          type: object
          additionalProperties: true

    PageMeta:
      type: object
      required: [page, size, returned]
      properties:
        page:
          type: integer
        size:
          type: integer
        returned:
          type: integer
        total:
          type: integer
          nullable: true
        has_next:
          type: boolean
          nullable: true

    DeviceListData:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DeviceRecord'
        page:
          $ref: '#/components/schemas/PageMeta'
        extensions:
          type: object
          additionalProperties: true

    DeviceListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/DeviceListData'

    DeviceDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/DeviceRecord'

    ActionResult:
      type: object
      required: [action, target]
      properties:
        action:
          type: string
          example: approve_device
        target:
          type: string
          example: 07f5e...
        success:
          type: boolean
          default: true
        extensions:
          type: object
          additionalProperties: true

    ActionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/ActionResult'

    RefreshTokenData:
      type: object
      required: [uuid, token]
      properties:
        uuid:
          type: string
        token:
          type: string
        rotated_at:
          type: string
          format: date-time
          nullable: true
        extensions:
          type: object
          additionalProperties: true

    RefreshTokenResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/RefreshTokenData'

    MetricPoint:
      type: object
      required: [ts, value, type]
      properties:
        ts:
          type: integer
          format: int64
          description: Unix 时间戳（毫秒）。
        value:
          type: number
          format: float
        type:
          $ref: '#/components/schemas/MetricType'
        extensions:
          type: object
          additionalProperties: true

    MetricsData:
      type: object
      required: [uuid, points]
      properties:
        uuid:
          type: string
        range:
          type: string
          nullable: true
          enum: [1h, 6h, 24h, 7d, all]
        start_ms:
          type: integer
          format: int64
          nullable: true
        end_ms:
          type: integer
          format: int64
          nullable: true
        points:
          type: array
          items:
            $ref: '#/components/schemas/MetricPoint'
        extensions:
          type: object
          additionalProperties: true

    MetricsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/MetricsData'

    User:
      type: object
      required: [username, permission, created_at]
      properties:
        username:
          type: string
        permission:
          $ref: '#/components/schemas/PermissionType'
        created_at:
          type: string
          format: date-time
        extensions:
          type: object
          additionalProperties: true

    UserListData:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        extensions:
          type: object
          additionalProperties: true

    UserListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/UserListData'

    UpdatePermissionRequest:
      type: object
      required: [permission]
      properties:
        permission:
          $ref: '#/components/schemas/PermissionType'
        extensions:
          type: object
          additionalProperties: true